requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.0
django-admin-rangefilter==0.13.3
django-filter==25.1
django-import-export==4.3.8
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
